Alan Brennan suggestion for change to Readme
see line 36 "Alan Brennan comment: I just wanted to practice making a suggested change to this brilliant and clear readme as if this was an actual suggested change in text - what i would recommend would be having the data sources in a more logical sequence for the reader.  i'd presumed you would say - they key data source is parkrun data which has the key variable ...  And then In addition to this we obtained data on blah - using source blah, and then data on blah2 using source blah2.  This way someone new to the paper and seeing the readme for the first time can follow and get a grip of the method. Fantastic work."